requirements.txt
flask
openai
pydantic
python-dotenv
